requirements.txt
pydantic>=2.0
fastapi>=0.100.0
uvicorn>=0.23.0
pycountry>=2.0
faker>=18.0.0
kafka-python>=2.0.0
python-dotenv>=0.19.0